# swashbuckle-dotnet7/docs/swagger.yaml
openapi: 3.0.1
info:
  title: dotnet7-swagger
  version: v1
paths:
  /WeatherForecast:
    get:
      tags:
        - WeatherForecast
      summary: Get all weather forecast.
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeatherForecast'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeatherForecast'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeatherForecast'
    post:
      tags:
        - WeatherForecast
      summary: Create weather forecast.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWeatherForecast'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateWeatherForecast'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateWeatherForecast'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateWeatherForecast'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateWeatherForecast'
  '/WeatherForecast/{type}':
    get:
      tags:
        - WeatherForecast
      summary: Get weather forecast by type.
      parameters:
        - name: type
          in: path
          description: Type of weather
          required: true
          style: simple
          schema:
            type: string
          example: Freezing
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateWeatherForecast'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMsg'
components:
  schemas:
    CreateWeatherForecast:
      type: object
      properties:
        date:
          type: string
          format: date
        temperatureC:
          type: integer
          format: int32
        summary:
          type: string
          nullable: true
      additionalProperties: false
    ErrorMsg:
      type: object
      properties:
        message:
          type: string
          nullable: true
      additionalProperties: false
    WeatherForecast:
      type: object
      properties:
        id:
          type: integer
          format: int32
        date:
          type: string
          format: date
        temperatureC:
          type: integer
          format: int32
        summary:
          type: string
          nullable: true
      additionalProperties: false